Move baseindexer test that uses win_type
