+++ pandas/tests/window/test_base_indexer.py
@@ -73,17 +73,6 @@ def test_indexer_accepts_rolling_args():
     tm.assert_frame_equal(result, expected)
 
 
-def test_win_type_not_implemented():
-    class CustomIndexer(BaseIndexer):
-        def get_window_bounds(self, num_values, min_periods, center, closed):
-            return np.array([0, 1]), np.array([1, 2])
-
-    df = DataFrame({"values": range(2)})
-    indexer = CustomIndexer()
-    with pytest.raises(NotImplementedError, match="BaseIndexer subclasses not"):
-        df.rolling(indexer, win_type="boxcar")
-
-
 @pytest.mark.parametrize("constructor", [Series, DataFrame])
 @pytest.mark.parametrize(
     "func,np_func,expected,np_kwargs",
